[refactor] Split scheduler.h into separate files per scheduler class
Move each scheduler implementation into its own header under
internal/scheduler/ for better modularity. Eliminate the
WorkSharingThreadPoolBase template by duplicating into standalone
WorkSharingThreadPool and PriorityThreadPool classes.
